Extract showing biometric prompt as an extension function
